feat; #144 Drawer and detail components created for Flight Search
