+++ src/app/flight/search-results/components/flight-detail.tsx
@@ -1,0 +1,8 @@
+/* eslint-disable prettier/prettier */
+import { useQueryStates } from 'nuqs'
+
+const FlightDetailsSearch = () => {
+  return <div className='text-xs text-blue-500'>Uçuş Detayları</div>
+}
+
+export { FlightDetailsSearch }
